refactor Report Class
- split into smaller ReportTypes (response, log, syscheck and review)
- Reports are not instantiated directly but only through a factory
