+++ backend/src/data-collection/ReportType.php
@@ -3,8 +3,8 @@
 declare(strict_types=1);
 
 enum ReportType: string {
-  case SYSCHECK = 'sys-check';
-  case RESPONSE = 'response';
   case LOG = 'log';
+  case RESPONSE = 'response';
   case REVIEW = 'review';
+  case SYSCHECK = 'sys-check';
 }
