+++ backend/test/unit/workspace/ReportTest.php
@@ -330,12 +330,12 @@ final class ReportTest extends TestCase {
     $this->reportFormat = ReportFormat::CSV;
 
     // Act
-    $report = new Report($this->workspaceId, $this->dataIds, $this->reportType, $this->reportFormat);
+    $report = new LogReportOutput($this->workspaceId, $this->dataIds, $this->reportFormat);
 
     // Assert
     parent::assertEquals($this->workspaceId, $report->getWorkspaceId());
     parent::assertEquals($this->dataIds, $report->getDataIds());
-    parent::assertEquals($this->reportType, $report->getType());
+    parent::assertEquals(get_class($report), LogReportOutput::class);
     parent::assertEquals($this->reportFormat, $report->getFormat());
   }
 
@@ -352,7 +352,7 @@ final class ReportTest extends TestCase {
       "\"sample_group\";\"sample_user\";\"xxx\";\"BOOKLET.SAMPLE-1\";\"\";\"\";\"1627545600\";sample log entry";
 
     // Act
-    $report = new Report($this->workspaceId, $this->dataIds, $this->reportType, $this->reportFormat);
+    $report = new LogReportOutput($this->workspaceId, $this->dataIds, $this->reportFormat);
     $report->setAdminDAOInstance($this->adminDaoMock);
     $generationSuccess = $report->generate();
 
@@ -370,7 +370,7 @@ final class ReportTest extends TestCase {
     $expectedLogsJsonReportData = self::LOGS;
 
     // Act
-    $report = new Report($this->workspaceId, $this->dataIds, $this->reportType, $this->reportFormat);
+    $report = new LogReportOutput($this->workspaceId, $this->dataIds, $this->reportFormat);
     $report->setAdminDAOInstance($this->adminDaoMock);
     $generationSuccess = $report->generate();
 
@@ -388,12 +388,7 @@ final class ReportTest extends TestCase {
     $this->adminDaoMock->allows('getLogReportData')->andReturn([]);
 
     // Act
-    $report = new Report(
-      $this->workspaceId,
-      $this->dataIds,
-      ReportType::LOG,
-      $reportFormat
-    );
+    $report = new LogReportOutput($this->workspaceId, $this->dataIds, $reportFormat);
     $report->setAdminDAOInstance($this->adminDaoMock);
     $generationSuccess = $report->generate();
 
@@ -418,7 +413,7 @@ final class ReportTest extends TestCase {
       '"sämple_group";"sämple_user";"xxx";"BOOKLET.SAMPLE-2";"UNIT.SÄMPLE";"";"""{\""name\"":\""S\u00e4m S\u00e4mple\"",\""age\"":42}""";""';
 
     // Act
-    $report = new Report($this->workspaceId, $this->dataIds, $this->reportType, $this->reportFormat);
+    $report = new ResponseReportOutput($this->workspaceId, $this->dataIds, $this->reportFormat);
     $report->setAdminDAOInstance($this->adminDaoMock);
     $generationSuccess = $report->generate();
 
@@ -436,7 +431,7 @@ final class ReportTest extends TestCase {
     $expectedResponsesJsonReportData = self::RESPONSES;
 
     // Act
-    $report = new Report($this->workspaceId, $this->dataIds, $this->reportType, $this->reportFormat);
+    $report = new ResponseReportOutput($this->workspaceId, $this->dataIds, $this->reportFormat);
     $report->setAdminDAOInstance($this->adminDaoMock);
     $generationSuccess = $report->generate();
 
@@ -454,12 +449,7 @@ final class ReportTest extends TestCase {
     $this->adminDaoMock->allows('getResponseReportData')->andReturn([]);
 
     // Act
-    $report = new Report(
-      $this->workspaceId,
-      $this->dataIds,
-      ReportType::RESPONSE,
-      $reportFormat
-    );
+    $report = new ResponseReportOutput($this->workspaceId, $this->dataIds, $reportFormat);
     $report->setAdminDAOInstance($this->adminDaoMock);
     $generationSuccess = $report->generate();
 
@@ -497,7 +487,7 @@ final class ReportTest extends TestCase {
         "\"sample_group\";\"sample_user\";\"xxx\";\"BOOKLET.SAMPLE-1\";\"\";\"1\";\"2021-07-29 10:00:00\";\"sample booklet review\"";
 
     // Act
-    $report = new Report($this->workspaceId, $this->dataIds, $this->reportType, $this->reportFormat);
+    $report = new ReviewReportOutput($this->workspaceId, $this->dataIds, $this->reportFormat);
     $report->setAdminDAOInstance($this->adminDaoMock);
     $generationSuccess = $report->generate($useNewVersion);
 
@@ -532,7 +522,7 @@ final class ReportTest extends TestCase {
         "\"sample_group\";\"sample_user\";\"xxx\";\"BOOKLET.SAMPLE-1\";\"\";\"1\";\"X\";\"X\";\"X\";\"2021-07-29 10:00:00\";\"sample booklet review\"";
 
     // Act
-    $report = new Report($this->workspaceId, $this->dataIds, $this->reportType, $this->reportFormat);
+    $report = new ReviewReportOutput($this->workspaceId, $this->dataIds, $this->reportFormat);
     $report->setAdminDAOInstance($this->adminDaoMock);
     $generationSuccess = $report->generate($useNewVersion);
 
@@ -556,7 +546,7 @@ final class ReportTest extends TestCase {
     );
 
     // Act
-    $report = new Report($this->workspaceId, $this->dataIds, $this->reportType, $this->reportFormat);
+    $report = new ReviewReportOutput($this->workspaceId, $this->dataIds, $this->reportFormat);
     $report->setAdminDAOInstance($this->adminDaoMock);
     $generationSuccess = $report->generate();
 
@@ -640,7 +630,7 @@ final class ReportTest extends TestCase {
     }
 
     // Act
-    $report = new Report($this->workspaceId, $this->dataIds, $this->reportType, $this->reportFormat);
+    $report = new ReviewReportOutput($this->workspaceId, $this->dataIds, $this->reportFormat);
     $report->setAdminDAOInstance($this->adminDaoMock);
     $generationSuccess = $report->generate($useNewVersion);
 
@@ -658,12 +648,7 @@ final class ReportTest extends TestCase {
     $this->adminDaoMock->allows('getReviewReportData')->andReturn([]);
 
     // Act
-    $report = new Report(
-      $this->workspaceId,
-      $this->dataIds,
-      ReportType::REVIEW,
-      $reportFormat,
-    );
+    $report = new ReviewReportOutput($this->workspaceId, $this->dataIds, $reportFormat);
     $report->setAdminDAOInstance($this->adminDaoMock);
     $generationSuccess = $report->generate();
 
@@ -688,7 +673,7 @@ final class ReportTest extends TestCase {
       "\"SAMPLE SYS-CHECK REPORT\";\"SYSCHECK.SAMPLE\";\"An example SysCheck definition\";\"\";\"" . filemtime(self::SYS_CHECK_SAMPLE_DATA_FILE) . "\";\"" . TimeStamp::toSQLFormat(filemtime(self::SYS_CHECK_SAMPLE_DATA_FILE)) . "\";\"" . basename(self::SYS_CHECK_SAMPLE_DATA_FILE) . "\";\"Linux\";\"x86_64\";\"1680 x 1050\";\"Chrome\";\"1\";\"Chromium PDF Plugin, Chromium PDF Viewer\";\"en-US\";\"79\";\"amd64\";\"8\";\"1680 x 914\";\"75.72 Mbit/s\";\"8.19 kbit/s\";\"good\";\"2.84 Mbit/s\";\"8.19 kbit/s\";\"good\";\"good\";\"100\";\"4g\";\"1.45\";\"Sam Sample\";\"Harvy Dent\";\"Because.\";\"1\";\"Radio Gaga\";\"1594.295166015625\"";
 
     // Act
-    $report = new Report($this->workspaceId, $this->dataIds, $this->reportType, $this->reportFormat);
+    $report = new SysCheckReportOutput($this->workspaceId, $this->dataIds, $this->reportFormat);
     $report->setSysChecksFolderInstance($this->sysChecksFolderMock);
     $generationSuccess = $report->generate();
 
@@ -723,7 +708,7 @@ final class ReportTest extends TestCase {
     ];
 
     // Act
-    $report = new Report($this->workspaceId, $this->dataIds, $this->reportType, $this->reportFormat);
+    $report = new SysCheckReportOutput($this->workspaceId, $this->dataIds, $this->reportFormat);
     $report->setSysChecksFolderInstance($this->sysChecksFolderMock);
     $generationSuccess = $report->generate();
 
@@ -743,12 +728,7 @@ final class ReportTest extends TestCase {
       ->andReturn([]);
 
     // Act
-    $report = new Report(
-      $this->workspaceId,
-      $this->dataIds,
-      ReportType::SYSCHECK,
-      $reportFormat
-    );
+    $report = new SysCheckReportOutput($this->workspaceId, $this->dataIds, $reportFormat);
     $report->setSysChecksFolderInstance($this->sysChecksFolderMock);
     $generationSuccess = $report->generate();
